mq_connect module, db connector func
